Fixed dragging down issue when sorting vertically (#47)
* Fixed dragging down issue on vertical sorting

The sorting worked fine on horizontal sorted elements. But when you wanted to sort items vertically (for example in a list), wrong element was selected when you was dragging down. You had to drag below the desired position to get the desired result. I changed the code so that you will select the next item, and not the current one when dragging down.

* Update index.js